added getBalance method
