+++ Sources/SolanaSwift/Model/SolanaSDK+Models.swift
@@ -35,11 +35,6 @@ protocol AnyDecodableInitializable {
 public struct Rpc<T: Decodable>: Decodable {
     public let context: Context
     public let value: T
-    
-    init(_ v: AnyDecodable) {
-        self.context = Context(slot: 1)
-        fatalError()
-    }
 }
 public struct Context: Decodable {
     public let slot: UInt64
